+++ uniorthocrin/database/migrations/2025_06_13_142610_create_campaign_posts_table.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('campaign_posts', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('campaign_id')->constrained('campaigns')->onDelete('cascade');
-            $table->string('name');
-            $table->text('description')->nullable();
-            $table->enum('type', ['feeds', 'stories_mg_sp', 'stories_df_es'])->default('feeds');
-            $table->foreignId('campaign_post_file_id')->nullable()->constrained('files')->onDelete('set null');
-            $table->enum('status', ['active', 'inactive'])->default('active');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('campaign_posts')) {
+            Schema::create('campaign_posts', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('campaign_id')->constrained('campaigns')->onDelete('cascade');
+                $table->string('name');
+                $table->text('description')->nullable();
+                $table->enum('type', ['feeds', 'stories_mg_sp', 'stories_df_es'])->default('feeds');
+                $table->foreignId('campaign_post_file_id')->nullable()->constrained('files')->onDelete('set null');
+                $table->enum('status', ['active', 'inactive'])->default('active');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
