+++ uniorthocrin/database/migrations/2025_06_13_142611_create_campaign_videos_table.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('campaign_videos', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('campaign_id')->constrained('campaigns')->onDelete('cascade');
-            $table->string('name');
-            $table->text('description')->nullable();
-            $table->enum('type', ['reels', 'marketing_campaigns'])->default('reels');
-            $table->foreignId('campaign_video_file_id')->nullable()->constrained('files')->onDelete('set null');
-            $table->enum('status', ['active', 'inactive'])->default('active');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('campaign_videos')) {
+            Schema::create('campaign_videos', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('campaign_id')->constrained('campaigns')->onDelete('cascade');
+                $table->string('name');
+                $table->text('description')->nullable();
+                $table->enum('type', ['reels', 'marketing_campaigns'])->default('reels');
+                $table->foreignId('campaign_video_file_id')->nullable()->constrained('files')->onDelete('set null');
+                $table->enum('status', ['active', 'inactive'])->default('active');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
